Work on Abilities and trying to fix sprite appearances

+++ modular_skyrat/modules/necromorphs/code/controllers/subsystems/necromorph.dm
@@ -35,6 +35,7 @@ SUBSYSTEM_DEF(necromorph)
 /datum/controller/subsystem/necromorph/stat_entry(msg)
 	return ("Click to debug!")
 
+
 /datum/controller/subsystem/necromorph/proc/join_necroqueue(mob/dead/observer/eye/signal/M)
 	if (is_marker_master(M))
 		//The master may not queue. They can still possess things if really needed though
@@ -42,7 +43,7 @@ SUBSYSTEM_DEF(necromorph)
 	necroqueue |= M
 	remove_verb(M, /mob/dead/observer/eye/signal/proc/join_necroqueue)
 	add_verb(M, /mob/dead/observer/eye/signal/proc/leave_necroqueue)
-	to_chat(M, ("You are now in the necroqueue. When a necromorph vessel is available, you will be automatically placed in control of it. You can still manually posess necromorphs."))
+	to_chat(M, span_notice("You are now in the necroqueue. When a necromorph vessel is available, you will be automatically placed in control of it. You can still manually posess necromorphs."))
 
 
 
@@ -59,13 +60,55 @@ SUBSYSTEM_DEF(necromorph)
 		if (!M.client || !M.key)
 			continue	//Gotta be connected
 
+		//The id is used to check preferences
+		if (!(vessel_id in M.client.prefs.be_special_role))
+			continue	//They don't wanna be this kind of necro, skip them
 
 		//If we get here, they'll do.
 		vessel.key = M.key	//Move into the mob and delete the old
 		qdel(M)
-
+		//GLOB.unitologists.add_antagonist(vessel.mind)
 		return TRUE
 	return FALSE	//Return false if we failed to find anyone
+
+
+
+//Sighting handling
+//-----------------------
+/*
+	Keeping track of last known locations of live humans
+*/
+/datum/controller/subsystem/necromorph/proc/update_sighting(var/mob/living/AM, var/obj/structure/corruption_node/eye/spotter)
+	//We don't record dead mobs, remove them from this list
+	if (AM.stat == DEAD || QDELETED(AM))
+		sightings.Remove(AM)
+		return 0	//This is numerically zero, not just false
+
+	//If we've ever seen this mob before, get their data from the list
+	var/datum/sighting/S = sightings[AM]
+	if (!istype(S))
+		//If we get here, they're new
+		S = new()
+		S.thing = AM
+		sightings[AM] = S
+
+	var/last_time = S.last_time
+
+	S.last_time = world.time
+	S.last_location = get_turf(AM)
+
+	var/difference = S.last_time - last_time
+
+	//Message all the necros with a useable jumplink
+	if (istype(spotter))
+		if (difference >= spotter.minimum_notify_delay)
+			for (var/key in SSnecromorph.necromorph_players)
+				var/mob/M = SSnecromorph.necromorph_players[key]
+				to_chat(M, SPAN_WARNING("[AM] detected at [jumplink_public(M, S.last_location)]"))
+
+
+	//We will return the difference between last and current time. Eyes may do something with this
+	return (difference)
 
 
 
@@ -98,6 +141,24 @@ SUBSYSTEM_DEF(necromorph)
 		return SSnecromorph.marker
 
 
+//Updates the energy holders of all necromorph players, refreshing their spell list
+/datum/controller/subsystem/necromorph/proc/update_all_ability_lists(var/clear = FALSE)
+	for (var/key in GLOB.players)
+		if (!key)
+			continue
+		var/datum/player/P = GLOB.players[key]
+		var/datum/extension/psi_energy/PE = P.get_energy_extension()
+		if (PE)
+			PE.build_ability_list(clear)
+			to_chat(P.get_mob(), "<span class='necromarker'>The marker has awoken, new abilities are unlocked. Check your abilities menu!</span>")
+
+		var/mob/dead/observer/eye/signal/S = P.get_mob()
+		if (istype(S))
+			S.update_verbs()
+
+
+
+
 //Shard handling
 /datum/controller/subsystem/necromorph/proc/register_shard(var/obj/item/marker_shard/MS)
 	var/shardsbefore = shards.len
@@ -106,7 +167,7 @@ SUBSYSTEM_DEF(necromorph)
 
 	//When the number of shards in the world switches between zero and nonzero, we update ability lists
 	if (shardsbefore == 0)
-
+		update_all_ability_lists()
 
 
 /datum/controller/subsystem/necromorph/proc/unregister_shard(var/obj/item/marker_shard/MS)
@@ -116,4 +177,4 @@ SUBSYSTEM_DEF(necromorph)
 
 	//When the number of shards in the world switches between zero and nonzero, we update ability lists
 	if ((shardsbefore > 0) && (shards.len <= 0))
-
+		update_all_ability_lists()
